school/serializers.py
from rest_framework import serializers

from school.models import Course, Lesson


class CourseSerializer(serializers.ModelSerializer):  # ViewSet

    class Meta:
        model = Course
        fields = '__all__'


class LessonSerializer(serializers.ModelSerializer):  # Generics

    class Meta:
        model = Lesson
        fields = '__all__'


